mudando as parada pra float

+++ grafo.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void InsereLista(Vertice *grafo, int v1, int v2, int c, int f){
+void InsereLista(Vertice *grafo, int v1, int v2, float c, float f){
 	Celula *nova = (Celula*)malloc(sizeof(Celula));
 	nova->prox = grafo[v1].arestas;
 	nova->capacidade = c;
